refactor(ai): Split Slice C lib modules and add source size ratchet (Refs #4079)
Concern-split nine planning files above 500 physical lines into named
siblings and wire repo.ai_source_file_size (500-line shrink-only gate).
