+++ packages/colonizethis_ai/lib/src/planning/conquest_planner.dart
@@ -1,5 +1,3 @@
-import 'package:colonizethis_logic/order_suggestion_api.dart';
-
 import 'conquest_move_scoring_context.dart';
 import 'conquest_planner_destination_scoring.dart';
 import 'goal_manager.dart';
@@ -17,6 +15,7 @@ import 'planning_helpers.dart'
         isAtWarWithAnyGreatPower,
         minorAtWarPeaceTargetsWhere;
 import '../util/ai_random_utils.dart';
+import 'conquest_planner_stalled_fallback.dart';
 
 export 'conquest_planner_destination_scoring.dart'
     show conquestOldWorldArmyMoveScaledBonus, conquestNwInvadableArmyMoveBonus;
@@ -286,7 +285,7 @@ Orders runConquestArmyMovePlanner({
       _log.d('conquest army move nationId=${ctx.nationId} candidatesCount=0');
     }
     if (stalledExpansion) {
-      return _runStalledFrontierArmyMoveFallback(
+      return runStalledFrontierArmyMoveFallback(
         ctx: ctx,
         scoringCtx: stalledScoringCtx!,
         feedstockConquestTarget: feedstockConquestTarget,
@@ -305,7 +304,7 @@ Orders runConquestArmyMovePlanner({
       _log.d('conquest army move filtered empty nationId=${ctx.nationId}');
     }
     if (stalledExpansion) {
-      return _runStalledFrontierArmyMoveFallback(
+      return runStalledFrontierArmyMoveFallback(
         ctx: ctx,
         scoringCtx: stalledScoringCtx!,
         feedstockConquestTarget: feedstockConquestTarget,
@@ -357,7 +356,7 @@ Orders runConquestArmyMovePlanner({
     return ctx.orders;
   }
   if (stalledExpansion) {
-    return _applyStalledArmyMovesForAllFieldArmies(
+    return applyStalledArmyMovesForAllFieldArmies(
       ctx: ctx,
       scoringCtx: stalledScoringCtx!,
       filtered: scoringCandidates,
@@ -433,98 +432,3 @@ ArmyMoveOrder? selectFeedstockBiasedBestArmyMove({
   return best;
 }
 
-Orders _applyStalledArmyMovesForAllFieldArmies({
-  required PlannerContext ctx,
-  required ConquestMoveScoringContext scoringCtx,
-  required List<ArmyMoveOrder> filtered,
-  required String? feedstockConquestTarget,
-}) {
-  final armiesWithOrders = <String>{
-    for (final m
-        in ctx.orders.armyMoveOrdersByPlayerId[ctx.nationId] ?? const [])
-      m.armyId,
-  };
-  final byArmy = <String, List<ArmyMoveOrder>>{};
-  for (final move in filtered) {
-    if (armiesWithOrders.contains(move.armyId)) continue;
-    (byArmy[move.armyId] ??= []).add(move);
-  }
-  var result = ctx.orders;
-  for (final armyId in byArmy.keys.toList()..sort()) {
-    final candidates = byArmy[armyId]!;
-    final best = selectFeedstockBiasedBestArmyMove(
-      candidates: candidates,
-      feedstockConquestTarget: feedstockConquestTarget,
-      score: (move) => scoreArmyMoveDestination(scoringCtx, move),
-    );
-    if (best == null) continue;
-    if (_log.infoEnabled) {
-      _log.i(
-        'conquest army move stalled multi nationId=${ctx.nationId} '
-        'armyId=${best.armyId} destinationProvinceId=${best.destinationProvinceId}',
-      );
-    }
-    result = applyArmyMoveOrderForPlayer(result, ctx.nationId, best);
-    armiesWithOrders.add(best.armyId);
-  }
-  return result;
-}
-
-Orders _runStalledFrontierArmyMoveFallback({
-  required PlannerContext ctx,
-  required ConquestMoveScoringContext scoringCtx,
-  required String? feedstockConquestTarget,
-}) {
-  final playerOwnedFullProvinceIds = <String>{
-    for (final e in ctx.view.provincesById.entries)
-      if (e.value.ownerId == ctx.nationId) e.key,
-  };
-  final validator = IncrementalCandidateValidator.forPlayer(
-    game: ctx.game,
-    topology: ctx.topology,
-    playerId: ctx.nationId,
-    basePrefix: ctx.orders,
-    factionMembership: DiplomacyFactionMembership.from(ctx.game),
-    resolution: orderResolutionContextFromView(ctx.view, ctx.game),
-  );
-  final armiesWithOrders = <String>{
-    for (final m
-        in ctx.orders.armyMoveOrdersByPlayerId[ctx.nationId] ?? const [])
-      m.armyId,
-  };
-  final acceptedCandidates = <ArmyMoveOrder>[];
-  for (final army in ctx.game.worldState.armies) {
-    if (army.ownerId != ctx.nationId || army.isHomeArmy) continue;
-    if (armiesWithOrders.contains(army.id)) continue;
-    final destIds = armyMoveCandidateDestinationProvinceIds(
-      game: ctx.game,
-      topology: ctx.topology,
-      playerId: ctx.nationId,
-      army: army,
-      playerOwnedFullProvinceIds: playerOwnedFullProvinceIds,
-    );
-    for (final destinationProvinceId in destIds) {
-      final candidate = ArmyMoveOrder(
-        armyId: army.id,
-        destinationProvinceId: destinationProvinceId,
-      );
-      if (!validator.isArmyMoveAccepted(candidate)) continue;
-      acceptedCandidates.add(candidate);
-    }
-  }
-  final best = selectFeedstockBiasedBestArmyMove(
-    candidates: acceptedCandidates,
-    feedstockConquestTarget: feedstockConquestTarget,
-    score: (candidate) => scoreArmyMoveDestination(scoringCtx, candidate),
-  );
-  if (best == null) {
-    return ctx.orders;
-  }
-  if (_log.infoEnabled) {
-    _log.i(
-      'conquest army move stalled fallback nationId=${ctx.nationId} '
-      'armyId=${best.armyId} destinationProvinceId=${best.destinationProvinceId}',
-    );
-  }
-  return applyArmyMoveOrderForPlayer(ctx.orders, ctx.nationId, best);
-}
